Refactor pools and staking tests
